refactor(gatekeep): renomear config_loader para config_gatekeep e isolar dependencias da CI
